extracted a SuroGenerator class to handle the question generation easily
