+++ suro-generator/themes.sample.txt
@@ -1,0 +1,1 @@
+histoire
